requirements.txt
flask
flask_login
torch
matplotlib
werkzeug
Flask-SQLAlchemy
Flask-Login
bcrypt
